backend/pyproject.toml
[project]
name = "imvu-insight-backend"
version = "0.1.0"
description = "FastAPI backend for IMVU Insight"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.128.0",
	"uvicorn[standard]>=0.40.0",
	"pyyaml>=6.0.2",

	"apscheduler>=3.11.0",

	"python-jose>=3.5.0",
	"ecdsa>=0.19.1",
	"pyasn1>=0.6.1",
	"rsa>=4.9.1",

	# Data processing
	"pandas>=2.2.3",
	"numpy>=2.4.0",
	"lxml>=5.3.0",
	"python-multipart>=0.0.6",

	# Database (MySQL)
	"sqlalchemy[asyncio]>=2.0.36",
	"asyncmy>=0.2.10",

	# Optional: migrations (can be wired later)
	"alembic>=1.14.0",
]

# uv supports dependency groups (PEP 735)
[dependency-groups]
dev = [
	"pytest>=8.3.4",
	"pytest-asyncio>=0.25.0",
	"httpx>=0.28.1",
	"ruff>=0.8.4",
	"black>=24.10.0",
	"mypy>=1.13.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "UP", "B"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
